==> 01.1_qn_dat_analysis.Rmd <==
---
title: "01.1_qn_dat_analysis"
author: "Puvvula"
date: "2023-03-16"
output: pdf_document
---

```{r}
library(pacman)
pacman::p_load(tidyverse, janitor, haven, reshape2)
path <- "~/Documents/MWAS_home/"
```


```{r}
#questionniare data
qn_dat<- read_csv(paste0(path, "pj_dat/qn_dat.csv"))|>
  mutate(stove = replace(stove, stove %in% c("NA_NA", "NA_lt_4hr"), NA))|>
  select(-regexer)

#creatinine adjusted values.
mom_pah_raw<- read_sas(paste0(path, "PAH_update/pahmetabolitewide3.sas7bdat")) |>
  select(c(1,2,6,7,9,13,14,16,20,21,23,27,28,30,34,35,37,41,42,44,48,49,51,55,56,58))|>
  clean_names() |>
  rename(creat_mg_dl="creatinine", group="participant_id")|>
  select(group, fluo2concentrationcr, nap1concentrationcr, nap2concentrationcr,
         phen1concentrationcr, phen4concentrationcr, phen9concentrationcr, 
         phen23concentrationcr, pyr1concentrationcr)|>
  rename_with(~str_sub(., end = -16), ends_with("concentrationcr"))|>
  mutate(sigma_lmwt_pah=fluo2+nap1+nap2+phen1+phen4+phen9+phen23)

summ<- mom_pah_raw |>
  pivot_longer(!c(group), names_to = "chemical", values_to = "conc")|>
  group_by(chemical) |>
  summarize(median_conc = round(median(conc),2),
            percentile_25 = round(quantile(conc, 0.25),2),
            percentile_75 = round(quantile(conc, 0.75),2))
```

```{r}
#creatinine adjusted PAH and questionnaire data
dt<- left_join(mom_pah_raw, dt_survey, by="group")  

qn_dat_pah <- dt |>
  mutate(mom_edu = if_else(mom_edu == "Bachelor's degree" | mom_edu == "Master's degree or above", "bach-or-abv", "lt-bach")) |>
  mutate(mom_race = if_else(mom_race == "Non-Hispanic White", "white", "non-white")) |>
  mutate(stove = replace(stove, stove %in% c("NA_NA", "NA_lt_4hr"), NA))|>
  mutate(stove = case_when(
    grepl("Gas", stove, ignore.case = TRUE) ~ "Gas",
    grepl("Electric", stove, ignore.case = TRUE) ~ "Electric",
    is.na(stove) ~ NA_character_,TRUE ~ stove)) |>
  rename(`1-Hydroxynapthalene` = "nap1",
         `2-Hydroxynapthalene` = "nap2",
         `2-Hydroxyfluorene` = "fluo2",
         `1-Hydroxyphenanthrene` = "phen1",
         `2,3-Hydroxyphenanthrene` = "phen23",
         `4-Hydroxyphenanthrene` = "phen4",
         `9-Hydroxyphenanthrene` = "phen9",
         `1-Hydroxypyrene` = "pyr1")|>
  select(-c(sigma_lmwt_pah, mom_age))|>
  pivot_longer(!c(1,10:16), names_to = "chemical", values_to = "conc")

dat<- qn_dat_pah
```


```{r}
#creatinine adjusted and log transformed values
dat<- read_csv(paste0(path, "pj_dat/pah_fin_impu.csv")) |>
  select(c(1,2:9,12:15,25,26)) |>
  mutate(stove = replace(stove, stove %in% c("NA_NA", "NA_lt_4hr"), NA))|>
  mutate(stove = case_when(
    grepl("Gas", stove, ignore.case = TRUE) ~ "Gas",
    grepl("Electric", stove, ignore.case = TRUE) ~ "Electric",
    is.na(stove) ~ NA_character_,TRUE ~ stove)) |>
  rename(`1-Hydroxynapthalene` = "nap1",
         `2-Hydroxynapthalene` = "nap2",
         `2-Hydroxyfluorene` = "fluo2",
         `1-Hydroxyphenanthrene` = "phen1",
         `2,3-Hydroxyphenanthrene` = "phen23",
         `4-Hydroxyphenanthrene` = "phen4",
         `9-Hydroxyphenanthrene` = "phen9",
         `1-Hydroxypyrene` = "pyr1")|>
  mutate(hh_inc_cat = if_else(is.na(hh_inc_cat), names(which.max(table(hh_inc_cat))), hh_inc_cat)) |>
  mutate(stove = if_else(is.na(stove), names(which.max(table(stove))), stove)) |>
  mutate(bbquse = if_else(is.na(bbquse), names(which.max(table(bbquse))), bbquse)) |>
  mutate(dist_majhw_cat = if_else(is.na(dist_majhw_cat), names(which.max(table(dist_majhw_cat))), dist_majhw_cat)) |>
  mutate_if(is.character, as.factor) |>
  rename(barbeque_use="bbquse", distance_to_highway="dist_majhw_cat", education="mom_edu_cat", income= "hh_inc_cat",
         stove_use="stove", race="race_cat")

```

#summary table
```{r}
summary_tbl <- function(data, categorical_vars, continuous_vars) {
  # initialize empty list to store results
  results_list <- list()
  
  # loop through each categorical variable
  for (cat_var in categorical_vars) {
    # loop through each level of the categorical variable
    for (cat_level in levels(data[[cat_var]])) {
      # subset the data to only include rows with the current categorical level
      cat_data <- data[data[[cat_var]] == cat_level, ]
      
      # initialize empty list to store summary statistics for each continuous variable
      cat_results_list <- list()
      
      # loop through each continuous variable
      for (cont_var in continuous_vars) {
        # compute summary statistics for the current continuous variable
        cont_summary <- as.data.frame(c(
          min = round(min(cat_data[[cont_var]]), 1),
          q1 = round(quantile(cat_data[[cont_var]], 0.25), 1),
          median = round(median(cat_data[[cont_var]]), 1),
          q3 = round(quantile(cat_data[[cont_var]], 0.75), 1),
          max = round(max(cat_data[[cont_var]]), 1),
          mean = round(mean(cat_data[[cont_var]]), 1),
          sd = round(sd(cat_data[[cont_var]]), 1),
          count = nrow(cat_data)
        ))
        
        # store the summary statistics in a named list
        cont_results <- list()
        cont_results$variable <- cont_var
        cont_results$level <- cat_level
        cont_results$min <- cont_summary[1, 1]
        cont_results$lower_quartile <- cont_summary[2, 1]
        cont_results$median <- cont_summary[3, 1]
        cont_results$upper_quartile <- cont_summary[4, 1]
        cont_results$max <- cont_summary[5, 1]
        cont_results$mean <- cont_summary[6, 1]
        cont_results$sd <- cont_summary[7, 1]
        cont_results$count <- cont_summary[8, 1]
        
        # add the summary statistics for the current continuous variable to the list
        cat_results_list[[cont_var]] <- cont_results
      }
      
      # combine the summary statistics for all continuous variables into a single data frame
      cat_results_df <- do.call(rbind, cat_results_list)
      
      # add the summary statistics for the current categorical level to the results list
      results_list[[paste(cat_var, cat_level, sep = "_")]] <- cat_results_df
    }
  }
  
  # combine the summary statistics for all categorical variables and levels into a single data frame
  results_df <- as.data.frame(do.call(rbind, results_list))
  rownames(results_df) <- NULL
  
  return(results_df)
}

x <- summary_tbl(dat, names(dat)[10:15], names(dat)[2:9])
write_csv(x, "~/Documents/MWAS_home/pj_results/grp_differences/summary_tbl.csv")
```

# Group difference testing - linear model, t-test and whitney-U test
```{r}
grp_diff_test_loop <- function(data, cat_vars, cont_vars) {
  result_df <- data.frame()  # initialize an empty dataframe for storing results

  for (cat_var in cat_vars) {
    for (cont_var in cont_vars) {
      # check if the continuous variable is numeric
      if (is.numeric(data[[cont_var]])) {
        # conduct the t-test and Mann Whitney U test
        t_test <- t.test(data[[cont_var]] ~ data[[cat_var]], conf.level=0.95)
        u_test <- wilcox.test(data[[cont_var]] ~ data[[cat_var]], conf.int = TRUE)
        lm_model <- lm(data[[cont_var]] ~ data[[cat_var]])
        
        # extract the relevant information from the t-test output
        t_stat <- t_test$statistic
        t_param <- t_test$parameter
        t_p_val <- t_test$p.value
        t_conf_int <- paste(round(t_test$conf.int[c(1,2)], 2), collapse = " - ")
        t_estimate_1 <- t_test$estimate[1]
        t_estimate_2 <- t_test$estimate[2]
        
        # extract the relevant information from the Mann Whitney U test output
        u_stat <- u_test$statistic
        u_p_val <- u_test$p.value
        u_conf_int <- paste(round(u_test$conf.int, 2), collapse = " - ")
        u_estimate <- u_test$estimate
        
        # extract the relevant information from the linear regression model
        lm_coefficients <- summary(lm_model)$coefficients
        lm_beta <- lm_coefficients[2]
        lm_conf_int <- paste(round(confint(lm_model)[2, ], 2), collapse = " - ")
        lm_p_val<- summary(lm_model)$coefficient[2,"Pr(>|t|)"]
        
        # create a row for the result dataframe
        result_row <- data.frame(cat_var = cat_var,
                                 cont_var = cont_var,
                                 t_stat = t_stat,
                                 t_df = t_param,
                                 t_p_val = t_p_val,
                                 t_conf_int = t_conf_int,
                                 t_mean_grp_1 = round(t_estimate_1, 2),
                                 t_mean_grp_2 = round(t_estimate_2, 2),
                                 u_stat = u_stat,
                                 u_p_val = u_p_val,
                                 u_conf_int = u_conf_int,
                                 u_median_diff = round(u_estimate, 2),
                                 lm_beta = lm_beta,
                                 lm_conf_int = paste(lm_conf_int),
                                 lm_p_val= lm_p_val)
        # add the row to the result dataframe
        result_df <- rbind(result_df, result_row)
      } else {
        # if the continuous variable is not numeric, print a warning message
        warning(paste0("Variable '", cont_var, "' is not numeric and will be skipped."))
      }
    }
  }
  return(result_df)
}
```

```{r}
#relevel variables - reference categories
dat$distance_to_highway <- relevel(dat$distance_to_highway, ref = "gt_1mile")
dat$education <- relevel(dat$education, ref = "bach-or-abv")
dat$race <- relevel(dat$race, ref = "white")
dat$income <- relevel(dat$income, ref = "abv_median")
dat$stove_use <- relevel(dat$stove_use, ref = "Electric")
dat$barbeque_use <- relevel(dat$barbeque_use, ref = "No")


# Results
x<- grp_diff_test_loop(dat, cont_vars = names(dat)[2:9], cat_vars = names(dat)[10:15])

write_csv(x, "~/Documents/MWAS_home/pj_result/grp_differences/grp_diff_results.csv")
```


#combine differences summary and t-test results
```{r}
dat<- read_csv("~/Documents/MWAS_home/pj_result/grp_differences/summary_tbl.csv") |>
  mutate_at(vars(1:3), factor)
dat_stats<- read_csv("~/Documents/MWAS_home/pj_result/grp_differences/grp_diff_results.csv") |>
  rename(pah = "cont_var", variable="cat_var")|>
  select(c(1,2,5,7,8,13,14,15)) |>
  mutate(percent_difference = round(abs(t_mean_grp_1 - t_mean_grp_2) / 
           ((t_mean_grp_1 + t_mean_grp_2) / 2) * 100, 1)) |>
  mutate_at(vars(1,2), factor)

# join group summaries with t-test results
dat_viz<- left_join(dat, dat_stats, by=c("pah", "variable")) |>
  mutate(x_dummy= "1") |>
  mutate(new_level = paste(level, count, sep = "-")) |>
  mutate(p_sig= if_else(lm_p_val< 0.05, 1, 2),
         pct_diff=if_else(percent_difference >10, 1, 2))

dat_viz$variable<- factor(dat_viz$variable, levels = c("race", "education", "income", 
                                           "distance_to_highway", "stove_use",
                                           "barbeque_use"))
######################################

```

#visualize and save plot
```{r}
cbbPalette <- c( "black", "grey65")

ggplot(dat_viz, aes(y = as.factor(new_level), x = median, xmin = pct_25, xmax = pct75)) + 
  geom_pointrange(aes(col = factor(p_sig)), position = position_dodge(width = 1), size = 0.5) +
  scale_colour_manual(values = cbbPalette) +
  xlab("") +
  ylab("") +
  theme_bw() +
  theme(
    axis.line = element_line(colour = "black"),
    axis.text = element_text(colour = "black", size = 8),
    strip.background = element_rect(fill = "transparent", colour = NA),
    strip.text = element_text(colour = "black", size = 8),
    strip.text.x = element_text(colour = "black", size = 6.5)
  ) +
  facet_grid(
    variable ~ factor(pah, levels = c(
      "1-Hydroxynapthalene", "2-Hydroxynapthalene", "2-Hydroxyfluorene",
      "1-Hydroxyphenanthrene", "2,3-Hydroxyphenanthrene", "4-Hydroxyphenanthrene",
      "9-Hydroxyphenanthrene", "1-Hydroxypyrene"
    )),
    scales = "free", switch = "x"
  ) +
  scale_y_discrete(labels = function(x) gsub("-", "\nn= ", x) %>% paste0("", .)) +
  geom_text(
    data = dat_viz %>%
      group_by(variable, pah) %>%
      mutate(beta = paste0("β=", round(lm_beta, 3)), ci = paste0("95% CI: ", lm_conf_int)),
    aes(x = Inf, y = -Inf, label = paste(beta, ci, sep = "\n")),
    hjust = 1.01, vjust = -0.1, size = 2.4, nudge_x = -0.2, nudge_y = -0.2, inherit.aes = FALSE
  ) +
  guides(color = FALSE)


ggsave("~/Documents/MWAS_home/pj_result/grp_differences/summary_plot.tiff", 
       width = 10,height = 7.3,
       dpi=300)
```

#correlation matrix
```{r}
covariates<- read_csv("~/Documents/MWAS_home/pj_dat/pah_fin_impu.csv") |>
  select(c(4:12)) |>
  rename(`1-Hydroxynapthalene` = "nap1",
         `2-Hydroxynapthalene` = "nap2",
         `2-Hydroxyfluorene` = "fluo2",
         `1-Hydroxyphenanthrene` = "phen1",
         `2,3-Hydroxyphenanthrene` = "phen23",
         `4-Hydroxyphenanthrene` = "phen4",
         `9-Hydroxyphenanthrene` = "phen9",
         `1-Hydroxypyrene` = "pyr1",
         `sigma-LMWT-PAH` = "sigma_lmwt_pah") 

corr <- cor(covariates, method = "pearson") |>
  melt()


ggplot(corr, aes(x = Var1, y = Var2, fill = value)) +
  geom_tile() +
  scale_fill_gradient2(low = "blue", mid = "white", high = "red", 
                       midpoint = 0, breaks = seq(-1, 1, 0.2)) +
  theme_minimal() +
  theme(axis.text.x = element_text(angle = 45, hjust = 1),
        strip.text = element_text(colour = "black", size = 8),
        strip.text.x = element_text(colour = "black", size = 6.5),
        strip.text.y = element_text(colour = "black", size = 6.5)) +
  labs(title = "", x = "", y = "") +
  geom_text(aes(label = round(value, 2)), color = "black", size = 3)+
  guides(fill=FALSE)

ggsave("~/Documents/MWAS_home/pj_results/grp_differences/corrmat.tiff", 
       width = 5,height = 5,
       dpi=300)
```
