==> 02_LC-MS_data_wrang.Rmd <==
---
title: "02.LC-MS data wrangling"
author: "Puvvula"
date: "2023-03-02"
output: pdf_document
---

```{r}
library(pacman)
pacman::p_load(tidyverse, janitor, haven, skimr, readr)
path <- "~/Documents/MWAS_home/pj_dat/lc_ms/"
```

c18 neg
```{r}
#LC-MS read values
c18neg<- read_tsv(paste0(path, "combat_c18neg.txt")) |>
  as.data.frame()

#subject_id - mapping
id_map<- read_tsv(paste0(path, "map_c18neg.txt")) |>
  as.data.frame() |>
  filter(substr(SampleID, 1, 1) == "A") |>
  filter(str_detect(SampleID, "M_2|B_2"))

# Mapped subject id's to the c18neg file
matched_names <- id_map$SampleID[match(names(c18neg), id_map$FileName)]
names(c18neg) <- ifelse(is.na(matched_names), names(c18neg), matched_names)
```

```{r}
# Extract mom or baby run
c18neg_start<- c18neg[c(1:10)]
mom_peaks<- grep("M_2$", names(c18neg), value = T)
df_mom_peak<- c18neg[,mom_peaks]

mom_dat<- cbind(c18neg_start, df_mom_peak) |>
  as.data.frame()

write_csv(mom_dat, "~/Documents/MWAS_home/pj_dat/lc_ms_c18neg_mom.csv")
```

==============================================================================
HILIC positive
```{r}
#LC-MS read values
hilicpos<- read_tsv(paste0(path, "combat_hilicpos.txt")) |>
  as.data.frame()

#subject_id - mapping
id_map<- read_tsv(paste0(path, "map_hilicpos.txt")) |>
  as.data.frame() |>
  filter(substr(SampleID, 1, 1) == "A") |>
  filter(str_detect(SampleID, "M_2|B_2"))

# Mapped subject id's to the hilicpos file
matched_names <- id_map$SampleID[match(names(hilicpos), id_map$FileName)]
names(hilicpos) <- ifelse(is.na(matched_names), names(hilicpos), matched_names)
```

```{r}
# Extract mom or baby run
hilicpos_start<- hilicpos[c(1:10)]
mom_peaks<- grep("M_2$", names(hilicpos), value = T)
df_mom_peak<- hilicpos[,mom_peaks]

mom_dat<- cbind(hilicpos_start, df_mom_peak) |>
  as.data.frame()

write_csv(mom_dat, "~/Documents/MWAS_home/pj_dat/lc_ms_hilicpos_mom.csv")
```

================================================================================
Annotation attempt - C18 Negative
https://hmdb.ca/spectra/ms/search
```{r}
library(xMSannotator)

mom_c18neg<- read_csv("~/Documents/MWAS_home/pj_dat/lc_ms_c18neg_mom.csv")
baby_c18neg<- read_csv("~/Documents/MWAS_home/pj_dat/lc_ms_c18neg_baby.csv")



```
