==> 03_MWAS.Rmd <==
---
title: "03_MWAS"
author: "Puvvula"
date: "2023-03-03"
output: pdf_document
---

```{r}
library(pacman)
pacman::p_load(tidyverse, janitor)
path <- "~/Documents/MWAS_home/pj_dat/"
```

#===============================================================================

## Select either mom or baby data from below

#MOM analysis
```{r}
covariates<- read_csv(paste0(path, "pah_fin_impu.csv"))  |>
  select(c(2:11,17,20,23,25,26)) |>
  rename(subject_id = "participant_id_mom",
         `1-Hydroxynapthalene` = "nap1",
         `2-Hydroxynapthalene` = "nap2",
         `2-Hydroxyfluorene` = "fluo2",
         `1-Hydroxyphenanthrene` = "phen1",
         `2,3-Hydroxyphenanthrene` = "phen23",
         `4-Hydroxyphenanthrene` = "phen4",
         `9-Hydroxyphenanthrene` = "phen9",
         `1-Hydroxypyrene` = "pyr1",
         `sigma-LMWT-PAH` = "sigma_lmwt_pah") |>
  mutate(across(where(is.character), factor))
```

#BABY analysis
```{r}
covariates<- read_csv(paste0(path, "pah_fin_impu.csv"))  |>
  select(c(2:11,18,20,21,23:26)) |>
  rename(subject_id = "participant_id_baby",
         `1-Hydroxynapthalene` = "nap1",
         `2-Hydroxynapthalene` = "nap2",
         `2-Hydroxyfluorene` = "fluo2",
         `1-Hydroxyphenanthrene` = "phen1",
         `2,3-Hydroxyphenanthrene` = "phen23",
         `4-Hydroxyphenanthrene` = "phen4",
         `9-Hydroxyphenanthrene` = "phen9",
         `1-Hydroxypyrene` = "pyr1",
         `sigma-LMWT-PAH` = "sigma_lmwt_pah") |>
  mutate(across(where(is.character), factor)) |>
  drop_na()
```

#===============================================================================
#===============================================================================
## Select only one dataset from this section

#mz and retention time - MOM - C18 Negative
```{r}
n_ft <- read_csv(paste0(path, "lc_ms_c18neg_mom.csv")) |>
  select(-c(3:10))|>
  mutate(across(3:last_col(), ~ log10(. + 1))) |>
  mutate(across(3:last_col(), ~ (. - mean(.)) / sd(.)))
```

#mz and retention time - MOM - hilic positive
```{r}
p_ft <- read_csv(paste0(path, "lc_ms_hilicpos_mom.csv")) |>
  select(-c(3:10))|>
  mutate(across(3:last_col(), ~ log10(. + 1))) |>
  mutate(across(3:last_col(), ~ (. - mean(.)) / sd(.)))
```

#mz and retention time - BABY - C18 Negative
```{r}
n_ft <- read_csv(paste0(path, "lc_ms_c18neg_baby.csv")) |>
  select(-c(3:10))|>
  mutate(across(3:last_col(), ~ log10(. + 1))) |>
  mutate(across(3:last_col(), ~ (. - mean(.)) / sd(.)))
```

#mz and retention time - BABY - hilic positive
```{r}
p_ft <- read_csv(paste0(path, "lc_ms_hilicpos_baby.csv")) |>
  select(-c(3:10))|>
  mutate(across(3:last_col(), ~ log10(. + 1))) |>
  mutate(across(3:last_col(), ~ (. - mean(.)) / sd(.)))
```


#===============================================================================
#align subject_id with the order from negative mz feature table
```{r shuffle FT columns}
matches <- c(1:2, match(covariates$subject_id, names(n_ft))) 
n_ft <- n_ft[matches]

#check order
match(names(n_ft)[-c(1:2)], covariates$subject_id)

#negative mode features
ft_neg<- n_ft |>
  unite(mz_rt, mz, time, sep = "_", remove = FALSE) |>
  mutate(mz_rt_mode = paste0(mz_rt, "_negative")) |>
  select(-c(mz, time, mz_rt)) |>
  select(mz_rt_mode, everything())
```

#align subject_id with the order from positive mz feature table
```{r}
matches <- c(1:2, match(covariates$subject_id, names(p_ft))) 
p_ft <- p_ft[matches]

#check order
match(names(p_ft)[-c(1:2)], covariates$subject_id)

#positive mode features
ft_pos<- p_ft |>
  unite(mz_rt, mz, time, sep = "_", remove = FALSE) |>
  mutate(mz_rt_mode = paste0(mz_rt, "_positive")) |>
  select(-c(mz, time, mz_rt)) |>
  select(mz_rt_mode, everything())
```

#combine negative and positive features
```{r}
feature_table<- bind_rows(ft_neg, ft_pos)

#save combined feature tables
write_csv(feature_table, paste0(path, "combined_ft/mom_ft.csv"))

write_csv(feature_table, paste0(path, "combined_ft/baby_ft.csv"))
```


#==============================================================================#
#==============================================================================#
#revised MWAS loop
#covariates for mom: mom_ppbmi + mom_age + pgtob + mom_edu_cat + race_cat
#covariates for baby: mom_ppbmi + mom_age + pgtob + mom_edu_cat + race_cat + totpg + sex
#==============================================================================#
```{r}
perform_PAH_MWAS <- function(feature_table, covariates, output_folder, mwas_file_name, cutoff_file_name) {
  result_list <- list() # create empty list to store results
  variables <- names(covariates[1:9]) # list of variables to analyze
  
  for (variable in variables) {
    current_results <- data.frame(matrix(nrow = nrow(feature_table), ncol = 4)) # create a new data frame for each variable
    names(current_results) <- c("Estimate", "Std. Error", "t value", "Pr(>|t|)")
    for (i in 1:nrow(feature_table)){
      metabolite <- unlist(feature_table[i,-c(1)]) # store the ith metabolite in new vector
      current_results[i,] <- summary(lm(metabolite ~ get(variable) + mom_ppbmi + mom_age + pgtob + mom_edu_cat + race_cat + totpg + sex,
                                        data=covariates))$coefficients[2,c(1:4)] # store regression results for ith metabolite
    }
    current_results$FDR <- p.adjust(current_results[[4]], method=c("fdr")) # fdr adjustment
    current_results <- cbind(feature_table[c(1)], current_results) # add mz and time
    current_results$Variable <- variable
    result_list[[variable]] <- current_results # add current results to list
  }
  
  PAH_MWAS_result <- do.call(rbind, result_list) # combine all results into one data frame
  
  # Perform additional step: Calculate FDR cutoff for each variable
  fdr_cutoff_values <- tapply(PAH_MWAS_result$FDR, PAH_MWAS_result$Variable, function(x) {
    sign_features <- sum(x < 0.2)
    count <- length(x)
    ((sign_features - 1) / (count - 1)) * 0.2
  })
  PAH_MWAS_result$fdr_cutoff <- fdr_cutoff_values[PAH_MWAS_result$Variable]
  
  # Calculate beta_dir variable
  PAH_MWAS_result$beta_dir <- case_when(
    PAH_MWAS_result$`Pr(>|t|)` < PAH_MWAS_result$fdr_cutoff & PAH_MWAS_result$Estimate > 0 ~ "positive-significant",
    PAH_MWAS_result$`Pr(>|t|)` < PAH_MWAS_result$fdr_cutoff & PAH_MWAS_result$Estimate < 0 ~ "negative-significant",
    PAH_MWAS_result$`Pr(>|t|)` > PAH_MWAS_result$fdr_cutoff & PAH_MWAS_result$Estimate > 0 ~ "positive-non_significant",
    PAH_MWAS_result$`Pr(>|t|)` > PAH_MWAS_result$fdr_cutoff & PAH_MWAS_result$Estimate < 0 ~ "negative-non_significant", TRUE ~ NA_character_)
  
  # Save final result to a file
  output_file <- paste0(output_folder, "/", mwas_file_name)
  write.csv(PAH_MWAS_result, file = output_file, row.names = FALSE)
    # Save unique values and corresponding FDR cutoffs to a separate file
  unique_variables <- unique(PAH_MWAS_result$Variable)
  fdr_cutoff_df <- data.frame(Variable = unique_variables, FDR_Cutoff = fdr_cutoff_values[unique_variables])
  fdr_output_file <- paste0(output_folder, "/", cutoff_file_name)
  write.csv(fdr_cutoff_df, file = fdr_output_file, row.names = FALSE)
  
  return(PAH_MWAS_result)
}
```

```{r}
result<- perform_PAH_MWAS(feature_table = feature_table, 
                          covariates = covariates,
                          output_folder = "~/Documents/MWAS_home/pj_result",
                          mwas_file_name="baby_mwas.csv",
                          cutoff_file_name="baby_fdr_coff.csv")
```

#==============================================================================#
#Volcano plot
#==============================================================================#
```{r}
ggplot(result, aes(x=(Estimate), y=-log10(`Pr(>|t|)`), color=beta_dir)) +
    geom_point(size=0.2) +
    scale_x_continuous(name="Adjusted beta coefficient per unit change in PAH level",
                       labels=scales::comma) +
    scale_y_continuous(name="-log10(P-value)", labels=scales::comma) +
    facet_wrap(~ factor(Variable, levels=c("1-Hydroxynapthalene", "2-Hydroxynapthalene", 
                                           "2-Hydroxyfluorene","1-Hydroxyphenanthrene",
                                           "2,3-Hydroxyphenanthrene","4-Hydroxyphenanthrene",
                                           "9-Hydroxyphenanthrene","sigma-LMWT-PAH", 
                                           "1-Hydroxypyrene")),
               ncol=3, nrow=3, scales="free") +
    theme_bw() +
    theme(panel.grid.major=element_blank(),
          panel.grid.minor=element_blank(),
          legend.position="bottom", # set legend position to bottom
          legend.box="horizontal") + 
    labs(color="Beta direction")+
    scale_color_manual(values = c("positive-significant" = "blue",
                                  "negative-significant" = "red",
                                  "positive-non_significant" = "grey",
                                  "negative-non_significant" = "grey"))

ggsave("~/Documents/MWAS_home/pj_result/baby_mwas.tiff",
       dpi=300, height = 8, width = 8)
```
