==> 05_pathway_import.Rmd <==
---
title: "05_pathway_import.Rmd"
author: "Puvvula"
date: "2023-03-14"
output: pdf_document
---

```{r}
  library(pacman)
pacman::p_load(tidyverse, janitor, readr)
```

#function to combine all the mimmichog import files
```{r}
combine_csv_files <- function(folder_location) {
  # Get list of CSV files in the folder
  file_list <- list.files(path = folder_location, pattern = "\\.csv$", full.names = TRUE)
  
  # Initialize an empty list to store data frames
  data_frames <- list()
  
  # Read and row bind each CSV file
  for (file in file_list) {
    # Extract chemical and mom_baby from file name
    file_name <- basename(file)
    file_parts <- strsplit(file_name, "_")[[1]]
    chemical <- file_parts[1]
    mom_baby <- gsub(".csv", "", file_parts[2], fixed = TRUE)
    
    # Read CSV file
    csv_data <- read.csv(file)
    
    # Add chemical and mom_baby as variables
    csv_data$chemical <- chemical
    csv_data$mom_baby <- mom_baby
    
    # Append the data frame to the list
    data_frames[[length(data_frames) + 1]] <- csv_data
  }
  
  # Combine all data frames into a single data frame
  combined_df <- do.call(rbind, data_frames)
  
  return(combined_df)
}
```

#merge data
```{r}
combined_data <- combine_csv_files("~/Documents/MWAS_home/pj_result/mummi_imp/p_3")
combined_data <- combined_data |> mutate(p_cutoff="p<0.5")

#export data
write_csv(combined_data, "~/Documents/MWAS_home/pj_result/mummi_import_combined_p3.csv")
```
